Task/SearchAPI
Added a SearchForm component to have the search bar and define the search view.
Added a index.js file to define the imports of the search folder.
Defined a search scene which would contain the search form and the search results.
Updated the index under the scenes folde to import the search.
Updated the MainNavBar to include the search button action.
Updated the MainNav to have a search option to call the Search Scene.
Updated the Search API file
Added a new search action which calls dispatch.