Task/SearchAPI
Added a search reducer.
Updated the search action to call the api.
Updated the api with the correct azure version. Need to clean up the code that does not use the helper funtion
Updated the Search scene with the redux connect code to get the correct store.
Only remaining part is to display the results.