Task/ProfilePage
Added a components folder to hold the profile page details.
Updated the user action to get the follow count.
Updated the user api to cal the follow count api.
Updated the user reducer to set the follow count
Updated the MainNav Bar to have the user profile button and navigate to that.
Added a main, info and index js files to profile component.
Added a profile scene.