Task/ProfilePage
Updated the user reducer with the car count, follow count and default cover Img.
Removed the set UserProfile from the Profile Scene
Updated the reference to the Cover image for the Profile page.
Updated the name to be displayed for the profile.
Updated the MainNavBar to redirect to the Profile scene.