Task/ProfilePage
Updated the profile page with the logic to get the follow count and the car count.
Updated the screen with the correct display of information.
Updated the Profile scene to user the correct reducers to get the details for the user and car.
Added new reducer methods to get the follow count and car count.