Task/CarRegistrationProcess
Updated the reducer with a new initial state object and added a new step for managing a different dispatch action.
Updated the profile scene by passing in the new navigator.
Updated the index under the Scene folder to have the AddCar js file.
Created a new scene for the Add Car process which defines the new routing required for the car registration flow.
Updated the main nav bar to have the add car scenario.
Updated the car garage with the navigation.
Added a new Success page for the end of the car registration process.
Added a Reg No Form for the car registration flow.
Added a Confirmation Detail screen for the car registration flow.
Updated the cars action with the loading and unloading status.