Task/CarProfileImage
Updated the Profile Scene to pass in the user object to the CarGarage component.
Updated the main nav class to pass in the route props.
Updated the CarGarageIcon with the image to be a button.
Updated the createCarGarageIcon function with the navigator push with correct route props such as the car  information and owner information.
Updated the Add Car Success page to have the correct icon.