Task/CarProfileImage
Updated the index page in scenes to have a new CarProfile scene.
Added a new Scene called CarProfile which displays a profile for a car.
Added two new components called Info and Gallery. Info will display the users information and car model information which the gallery will display all the images as a gallery.