Task/CarProfile
Updated the reducer to hold the car post images.
Updated the car profile to load the initial images.
Added backscene to the current view to load the car profile with navigation.
Updated the backscene to have the title be passed into the view.
Removed the car model details from the Info object.
Updated the gallery to be a component which could be used in other places of the app. Removed the scroll view and add a list view.
Updated the CarGarage to have the correct id passed into the key.
Added a fetch limit as a config.
Added the api call to get the image content from the backend service.
Updated the cars action to get the cars and then apply a results into a object type before dispatching into the reducer.