Task/CarProfileShowCase
Updated the CarProfile to have a tab view using a new scrollable-tab-view library.
Added a new action and api method to get the videos from the backend.
Updated the reducer to store the videos.
Updated the Gallery view to display the Videos.
The Car Profile Gallery view was updated with a Scroll View that displays a scroll view and a list view. 
Updated the ListVideo with the correct dimensions.