Task/CarProfileTimeline
Updated the car profile timeline with the comments featrure.
Added a new comments reducer to help with storing the comments and reflecting them immediately.
Updated the comments input and comments view to match the required designs.
Removed Comments input.
Added a new comments action to call the api service to create and get comments.
Removed any comments calls from the post action.
Updated the types to match the new reducer.
Updated the timeline scene to handle the logic for the comments.