Task/AddPostFunctionality
Added a new post reducer to hold the post data from one screen shift to another.
Removed a ddos code from Timeline.js to get user likes.
Updated the AddPost scene to have a navigator to render the correct view and added in the redux connect functionality to connect with the post reducer.
Broke the AddPost screen into multiple reusable screens between landing page and confirm page. The screens are TagsView, StatusView,GalleryView and CarSelectionView.
Updated the landing page to call the common views.
Updated the confirm page to have the final confirm screen before posting.
Updated the Reducer Types with the new types required for add post.
Updated the logic in the timeline action to add in a image or video and added in a common method to call from the addpost screens.
Updated the post action to update the add post store and reset the post data.