Task/FollowUnFollowFeature
Updated the cars action to have the follow and unfollow method.
Added two new types.
Consolidated update image and video content types to one.
Updated the car api to call the follow and unfollow api.
Updated the car profile info view to have stats data.
Updated the car profile scene to add in the type of content.
Updated the cars reducer to have the data of cars that are being browsed and to hold the content of that car. This gives granular level of data storage per each car viewed.