Task/FollowUnFollowFeature
Updated the GetStream api to have a getUserFollowing and CarFollowers method.
Updated the user action class to have a getUserFollowingFeeds method which gets the feeds the user follows and adds it to the user store.
Updated the user api to get the get stream token.
Updated the CarProfile to getUserFollowingFeeds, followCar and unFollow car.
Updated the user reducer to store the car id's the user is following.