Task/ViewCarProfileFromSearch
Updated the cars action to have a getCarbyId method and a new method to set the browsing car timeline.
Added a new type named Set Browsing Car
Updated the CarProfile class to distinguish between the car profile coming from search screen and garage screen.
Updated the seach car component to pass in the navigator.
Updated the search row to have a arrow to take the user to the car profile and this would call the navigator to push the screen to the car profile.
Updated the reducer to set the browsing car data in the browsingCars array.