Task/ViewCarProfileFromSearch
Updated the Car action to call the set browsing car dispatch method.
Updated the getstream method to pass the token to the headers.
Updated the gallery to read the correct uris.
Renamed the user to car owner in Post class.
Updated the CarProfile with some syntax errors.
Updated the timeline class to pass carOwner to getPost method.
Updated the cars reducer to handle the videos and images.