Task/GetRealFollowersData
Added the logic required to get the real followers data.
Updated the cars action to get followers from get stream.
Added a new type to set the followers count.
Updated the user cars api call.
Updated the CarProfileInfo view to get the details from a carStats object.
Updated the CarProfileScene to call the getCarFollowersCount data.
Updated the profile call to pass in the user id.
Updated the model in the cars reducer.