Task/VerifyCar
Added new type
Added new user action to call the verify car api.
Updated the user car api class to build the verify car request.
Added a new generic action button
Updated the style for the new action button
Updated the car profile info to have the action button and pass in the relevant functions.
Added in a verify folder and add in the intro, confirm and success screens which is supported by an index file.
Updated the MainNav class to have the VerifyCar scene
Added the code to CarProfile to check if the car is verified. This is commented off at hte moment.
Updated the cars reducer to have the verified flag.