Task/InviteFeature
Added RequestForm, RequestInvite, StatusView and index view for the invite scene.
Updated the RootNav code to trigger the invite scene.
Added designs for the invite screens.
Updated the GradientView to display full screen or only 95% of the screen leaving space for the tab bar.
Updated the Button index and styles with a new button.
Updated palette with new color scheme