Task/InviteFeature
Added invite actions, api and reducer to store the invite reference and status.
Added Asyncstorage to store and retrieve the invite reference from phone storage during start up.
Updated the logic in the views to submit the data to the backend,.
Utilised keyboard aware view for the screen so the screen moves with the keyboard poping up.